# src/main/java/com/metaverse/msme/msme_unit_details/controller/MsmeUnitDetailsController.java
package com.metaverse.msme.msme_unit_details.controller;

import com.metaverse.msme.common.ApplicationAPIResponse;
import com.metaverse.msme.msme_unit_details.service.*;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.validation.Valid;
import lombok.RequiredArgsConstructor;
import org.springframework.data.domain.Page;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.security.access.prepost.PreAuthorize;
import org.springframework.security.core.Authentication;
import org.springframework.security.core.context.SecurityContextHolder;
import org.springframework.web.bind.annotation.*;

import java.security.Principal;
import java.util.List;
import java.util.Map;

@RestController
@RequestMapping("/api/msme-unit")
@RequiredArgsConstructor
@Tag(name = "MSME Unit Details", description = "MSME Unit management and search APIs")
public class MsmeUnitDetailsController {

    private final MsmeUnitDetailsService msmeUnitDetailsService;

    @PutMapping("/update")
    @Operation(
            summary = "Update MSME unit details",
            description = "Update existing MSME unit information by unit ID"
    )
    @SecurityRequirement(name = "Bearer Authentication")
    @PreAuthorize("isAuthenticated()")
    public ResponseEntity<?> updateMsmeUnitDetails(
            @Parameter(description = "MSME Unit ID") @RequestParam(required = false) Long msmeUnitId,
            @RequestBody MsmeUnitDetailsDto request) {

        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
        String userId = authentication.getName();

        if (userId == null || "anonymousUser".equals(userId)) {
            ApplicationAPIResponse<String> response = ApplicationAPIResponse.<String>builder()
                    .success(false)
                    .message("User authentication required")
                    .code(401)
                    .build();
            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(response);
        }

        MsmeUnitDetailsDto updated = msmeUnitDetailsService.updateMsmeUnitDetails(msmeUnitId, request, userId);

        ApplicationAPIResponse<MsmeUnitDetailsDto> response = ApplicationAPIResponse.<MsmeUnitDetailsDto>builder()
                .data(updated)
                .success(true)
                .message("MSME unit details updated successfully")
                .code(200)
                .build();

        return ResponseEntity.ok(response);
    }

    @GetMapping("/{id}")
    @Operation(
            summary = "Get MSME unit by ID",
            description = "Retrieve MSME unit details by unit ID"
    )
    @SecurityRequirement(name = "Bearer Authentication")
    public ResponseEntity<?> getMsmeUnit(
            @Parameter(description = "MSME Unit ID") @PathVariable Long id) {

        MsmeUnitDetailsDto unitDetails = msmeUnitDetailsService.getMsmeUnitById(id);

        ApplicationAPIResponse<MsmeUnitDetailsDto> response = ApplicationAPIResponse.<MsmeUnitDetailsDto>builder()
                .data(unitDetails)
                .success(true)
                .message("MSME unit details retrieved successfully")
                .code(200)
                .build();

        return ResponseEntity.ok(response);
    }

    @PostMapping("/search")
    @Operation(
            summary = "Search MSME units",
            description = "Search MSME units with flexible criteria: districts, mandals, villages, unit name, or mobile number. " +
                    "Supports multiple combinations and partial matching for text fields.")
    public ResponseEntity<?> searchMsmeUnits(
            @Parameter(description = "Page number (0-based). Overrides body.page if provided.")
            @io.swagger.v3.oas.annotations.parameters.RequestBody(
                    description = "Search criteria. All fields are optional. Can combine multiple filters.",
                    required = true,
                    content = @Content(schema = @Schema(implementation = MsmeUnitSearchRequest.class)))
            @RequestBody MsmeUnitSearchRequest request) {

        Page<MsmeUnitSearchResponse> results = msmeUnitDetailsService.searchMsmeUnits(request);

        MsmeUnitSearchPageResponse pageResponse = MsmeUnitSearchPageResponse.builder()
                .content(results.getContent())
                .pageNumber(results.getNumber())
                .pageSize(results.getSize())
                .totalElements(results.getTotalElements())
                .totalPages(results.getTotalPages())
                .build();

        ApplicationAPIResponse<MsmeUnitSearchPageResponse> response =
                ApplicationAPIResponse.<MsmeUnitSearchPageResponse>builder()
                        .data(pageResponse)
                        .success(true)
                        .message(results.getTotalElements() + " units found")
                        .code(200)
                        .build();

        return ResponseEntity.ok(response);
    }


    @GetMapping(path = "/dashboard")
    public ResponseEntity<?> summaryOfMsmeUnits(Principal principal) {
        String userId = principal.getName();
        MsmeUnitSummaryResponse summary = msmeUnitDetailsService.summaryOfMsmeData();
        ApplicationAPIResponse<MsmeUnitSummaryResponse> response =
                ApplicationAPIResponse.<MsmeUnitSummaryResponse>builder()
                        .data(summary)
                        .success(true)
                        .message("MSME unit summary retrieved successfully")
                        .code(200)
                        .build();

        return ResponseEntity.ok(response);
    }

    @GetMapping(path = "/villages-progress/{district}")
    public ResponseEntity<?> summaryOfVillagesMsmeUnits(@PathVariable String district, @RequestParam String mandal, @RequestParam(required = false,defaultValue = "0") Integer page, @RequestParam(required = false,defaultValue = "10") Integer size) {
        Page<MsmeUnitSummaryResponse> summaryPage = msmeUnitDetailsService.summaryOfMsmeData(district, mandal, page, size);
        MsmeUnitSummaryPageResponse summary = MsmeUnitSummaryPageResponse.builder()
                .content(summaryPage.getContent())
                .pageNumber(summaryPage.getNumber())
                .pageSize(summaryPage.getSize())
                .totalElements(summaryPage.getTotalElements())
                .totalPages(summaryPage.getTotalPages())
                .build();

        ApplicationAPIResponse<MsmeUnitSummaryPageResponse> response =
                ApplicationAPIResponse.<MsmeUnitSummaryPageResponse>builder()
                        .data(summary)
                        .success(true)
                        .message("Village-progress MSME unit summaries retrieved successfully")
                        .code(200)
                        .build();

        return ResponseEntity.ok(response);
    }

    @PostMapping(path = "/duplicate-probability")
    public ResponseEntity<?> duplicateProbability(@RequestBody DuplicateProbabilityRequest request) {
        List<DuplicateProbability> probabilities = msmeUnitDetailsService.duplicateProbabilityCheck(
                request.getUnitName(), request.getOwnerName(), request.getDistrict(), request.getMandal(), request.getVillages());

        ApplicationAPIResponse<List<DuplicateProbability>> response =
                ApplicationAPIResponse.<List<DuplicateProbability>>builder()
                        .data(probabilities)
                        .success(true)
                        .message("Duplicate probability evaluated successfully")
                        .code(200)
                        .build();

        return ResponseEntity.ok(response);
    }

    @Operation(summary = "Mark as Duplicate",
            description = "This API marks an existing record as duplicate to prevent further processing or duplication conflicts.")
    @PutMapping(path = "/mark-duplicate")
    public ResponseEntity<?> markMsmeDuplicate(@RequestBody List<Long> msmeUnitId) {
        int updatedCount = msmeUnitDetailsService.markMsmeDuplicate(msmeUnitId);
        if (updatedCount > 0) {
         return ResponseEntity.ok(ApplicationAPIResponse.<Integer>builder().data(updatedCount).success(true).message("Marked duplicate msme's").code(200).build());
        }
        return ResponseEntity.ok(ApplicationAPIResponse.<Integer>builder().data(updatedCount).success(false).message("msme's can not find "+ msmeUnitId).code(400).build());
    }

    @Operation(
            summary = "MSME drilldown summary dashboard",
            description = "district is mandatory. If mandal is empty, returns mandal-wise counts for the district. If mandal is provided, returns village-wise counts for that mandal."
    )
    @io.swagger.v3.oas.annotations.parameters.RequestBody(
            required = true,
            content = @Content(schema = @Schema(implementation = DistrictMandalSummaryRequest.class))
    )
    @io.swagger.v3.oas.annotations.responses.ApiResponses(value = {
            @io.swagger.v3.oas.annotations.responses.ApiResponse(
                    responseCode = "200",
                    description = "MSME unit drilldown summary retrieved successfully"
            ),
            @io.swagger.v3.oas.annotations.responses.ApiResponse(
                    responseCode = "400",
                    description = "Invalid input"
            )
    })
    @PostMapping(path = "/summary/drilldown-dashboard")
    public ResponseEntity<?> summaryDrilldown(@Valid @RequestBody DistrictMandalSummaryRequest request) {
    Map<String, MsmeUnitSummaryResponse> summary = msmeUnitDetailsService.summaryOfMsmeDataDrilldown(
                request.getDistrict(),
                request.getMandal()
        );

        ApplicationAPIResponse<Map<String, MsmeUnitSummaryResponse>> response =
                ApplicationAPIResponse.<Map<String, MsmeUnitSummaryResponse>>builder()
                        .data(summary)
                        .success(true)
                        .message("MSME unit drilldown summary retrieved successfully")
                        .code(200)
                        .build();

        return ResponseEntity.ok(response);
    }

}